fix(schema-editor): use shared ColumnIcon to match SQL Editor
